requirements.txt
PyYAML~=6.0
Telethon~=1.23.0